update products settings
