+++ components/product-settings/index.ts
